requirements.txt
django
djangorestframework 
mlflow 
scikit-learn
pandas 
numpy 
gunicorn